requirements.txt
beautifulsoup4==4.6.0
certifi==2018.1.18
chardet==3.0.4
cssselect==1.0.3
defusedxml==0.5.0
Django==2.0.1
django-allauth==0.35.0
django-extensions==1.9.9
django-webpack-loader==0.5.0
feedparser==5.2.1
idna==2.6
lxml==4.1.1
mysqlclient==1.3.12
oauthlib==2.0.6
pyquery==1.4.0
python-json-logger==0.1.8
python3-openid==3.1.0
pytz==2017.3
requests==2.18.4
requests-oauthlib==0.8.0
six==1.11.0
typing==3.6.4
urllib3==1.22
